+++ Kernel/Output/HTML/LayoutTicket.pm
@@ -181,6 +181,7 @@ sub AgentQueueListOption {
     my $Class          = defined( $Param{Class} )          ? $Param{Class}          : '';
     my $SelectedIDRefArray = $Param{SelectedIDRefArray} || '';
     my $Multiple       = $Param{Multiple}                  ? 'multiple = "multiple"' : '';
+    my $TreeView       = $Param{TreeView}                  ? $Param{TreeView}        : 0;
     my $OptionTitle    = defined( $Param{OptionTitle} )    ? $Param{OptionTitle}     : 0;
     my $OnChangeSubmit = defined( $Param{OnChangeSubmit} ) ? $Param{OnChangeSubmit}  : '';
     if ($OnChangeSubmit) {
@@ -360,6 +361,10 @@ sub AgentQueueListOption {
         }
     }
     $Param{MoveQueuesStrg} .= "</select>\n";
+
+    if ($Param{TreeView}) {
+        $Param{MoveQueuesStrg} .= ' <a href="#" title="$Text{"Show Tree Selection"}" class="ShowTreeSelection">$Text{"Show Tree Selection"}</a>';
+    }
 
     return $Param{MoveQueuesStrg};
 }