+++ CHANGES.md
@@ -1,4 +1,6 @@
 #3.3.0.beta1 2013-??-??
+ - 2013-06-11 Added new feature to allow selecting services, queues, etc. from a JSTree view which is opened
+    within an overlay (agent and customer frontend).
  - 2013-06-10 Added new feature that set a ticket to be displayed as an even on a ticket dashboard widget.
  - 2013-06-10 Fixed bug#[5759](http://bugs.otrs.org/show_bug.cgi?id=5759) - Company ID is listed twice in SQL field list.
  - 2013-06-10 Fixed bug#[9366](http://bugs.otrs.org/show_bug.cgi?id=9366) - Order of involved agents in AgentTicketNote is incorrect.